app helpers refactor
